Profile Builder Helper and File Consolidation
- Created a new helper for profile building in universal-helpers
- Renamed game.py for better clarity
- Renamed many functions for clarity
- Removed redundant documentation
- Added section headers
- Removed unnecessary exports
- Consolidated communities.py via new helpers
- Removed two dead functions in communities.py
